Return int from func_ov004_020b8f78 so its one caller reads a whole word
The minigame results panel's play-again buttons never appeared and never
accepted a tap. The floor was not the panel's fill: func_ov004_020af27c,
the ROM's own slot 27, writes all six position halfwords correctly, and
(-128,48) (384,96) (128,224) are the OFF-SCREEN START of a slide-in.

The slide is in func_ov004_020aeb24, which also carries the stylus hit
test, and it never reached either. Its first statement is

    if (func_ov004_020b8f78(c + 0xf4) != 0) return;

and it declares that callee int while the callee was DEFINED returning u8.
mwccarm compiles the callee's tail to `ldrb r0, [r4, #0x124]` under either
spelling, which zero-extends into the whole of r0, so the ROM agrees with
itself and the byte gate is right to be silent. MSVC emits
`mov al, [esi+124h]` and leaves EAX's top 24 bits holding 0xFFFFFF from the
two `sub eax,1` compares in the switch above, so the caller's `test eax,eax`
read 0xFFFFFF00 and took the early return on every frame.

int is the type the ROM's ABI actually delivers here, and the sibling
predicate at 0x020b8f18 already carries it. Verified unchanged: MATCH at
2004/b56, 1.2/base, 1.2/sp2 and 1.2/sp2p3, and linkcheck VERIFIED with 0
blind relocs.

Measured on scene 384 with the panel raised through slot 27: the three
buttons slide to (128,48) (128,96) (128,144) over sixteen frames, OAM
engine B goes from 0 placed to 115 placed, and a scripted tap at (128,48)
moves the selection off -1, steps the confirm counter to 0x10000 and takes
the panel down through slot 28. Same on seven other minigame scenes, which
is expected: this is dScMgBase_c framework code every class shares.

+++ src/func_ov004_020b8f78.cpp
@@ -15,7 +15,21 @@ u8 DecIfAbove0_Byte(u8 *p);
 void func_ov004_020b91fc(char *c);
 }
 
-extern "C" u8 func_ov004_020b8f78(char *c)
+/* THE RETURN TYPE IS int AND IT IS LOAD-BEARING ON THE HOST, run mg14 lane
+ * RESULTS. The tail is `return *(u8 *)(c + 0x124);` and mwccarm compiles that
+ * to `ldrb r0, [r4, #0x124]` whichever type this says, so the ROM bytes are
+ * identical either way -- verified MATCH at 2004/b56 and the 1.2 trio with
+ * both spellings. MSVC is not: with a u8 return it emits `mov al, [esi+124h]`
+ * and leaves the top 24 bits of EAX holding whatever the switch above left
+ * there (0xFFFFFF on the state-0 path, from the two `sub eax,1` compares), and
+ * the ONE caller -- func_ov004_020aeb24, the results panel's tick -- declares
+ * this int and tests the full register. So the minigame results panel took its
+ * early return on every frame, the three ApproachLinear2 slides never ran, and
+ * the play-again buttons sat at their off-screen seeds forever.
+ *
+ * That is the whole defect behind "the play-again row never appears". Do not
+ * narrow this back to u8 to match the field it returns. */
+extern "C" int func_ov004_020b8f78(char *c)
 {
     u8 st = *(u8 *)(c + 0x124);
     switch (st) {
